requirements.txt
Django==2.2
django-cors-headers==3.4.0
django-redis==4.12.1
djangorestframework==3.11.0
pytz==2020.1
redis==3.5.3
StackAPI==0.1.12

